+++ cargo-lock/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- MSRV 1.58 to support more ported-from-cargo code ([#559])
+
 ## 8.0.3 (2022-11-30)
 ### Fixed
 - Encoding inconsistency when there's only one registry for all packages ([#767])